+++ monai/data/__init__.py
@@ -16,12 +16,12 @@ from .dataset import (
     CacheDataset,
     CacheNTransDataset,
     Dataset,
-    IterableDataset,
     LMDBDataset,
     PersistentDataset,
     SmartCacheDataset,
     ZipDataset,
 )
+from .iterable_dataset import IterableDataset
 from .decathlon_datalist import load_decathlon_datalist, load_decathlon_properties
 from .grid_dataset import *
 from .image_reader import *
